chaining: parse 64-bit CLI options with atoll, not atoi
The -g/--maxGapLength, -d/--chainGapOpen and -e/--chainGapExtend options
back int64_t fields but were parsed with atoi(), which returns a 32-bit
int.  Any value above INT_MAX (e.g. -g 3000000000) silently wraps to a
negative number, after which the max_gap_length break/remove tests reject
every link and chaining becomes a no-op -- with no diagnostic.

Switch the three int64 options to atoll().  Defaults (1000000/5000/1) are
in range so behavior is byte-identical there; values above INT_MAX are now
honored.  Verified: a 2.5e9-gap pair chains at -g 3000000000 and not at
-g 2000000000 (boundary straddles INT_MAX, previously impossible).

+++ impl/paf_chain.c
@@ -86,16 +86,16 @@ int paffy_chain_main(int argc, char *argv[]) {
                 outputFile = optarg;
                 break;
             case 'g':
-                max_gap_length = atoi(optarg);
+                max_gap_length = atoll(optarg);
                 break;
             case 't':
                 percentage_to_trim = atof(optarg);
                 break;
             case 'd':
-                chain_gap_open = atoi(optarg);
+                chain_gap_open = atoll(optarg);
                 break;
             case 'e':
-                chain_gap_extend = atoi(optarg);
+                chain_gap_extend = atoll(optarg);
                 break;
             case 'h':
                 usage();
